Refactor fees and trade summary
